테스트 어노테이션 및 마이그레이션 공통화
