+++ src/test/java/hello/ApplicationTest.java
@@ -1,35 +1,40 @@
 package hello;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 import org.junit.Test;
+import org.junit.Before;
+import org.junit.After;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.TestPropertySource;
+
+import hello.Migration;
 
 // test시 사용할 properties파일을 선택한다
 @TestPropertySource(locations ="classpath:/application-test.properties")
 @RunWith(SpringRunner.class)
 @SpringBootTest
-@AutoConfigureMockMvc
 public class ApplicationTest
 {
-    private static final Logger LOG = LoggerFactory.getLogger(ApplicationTest.class);
-
     @Autowired
-    private MockMvc mockMvc;
+    private Migration migration;
+
+    @Before
+    public void setup()
+    {
+        migration.testUp();
+    }
 
     @Test
-    public void shouldReturnDefaultMessage() throws Exception {
-        this.mockMvc.perform(get("/login")).andDo(print()).andExpect(status().isOk());
+    public void mainTest() throws Exception
+    {
+    }
+
+    @After
+    public void tearDown()
+    {
+        // migration.testClean();
     }
 }
