+++ src/test/java/hello/Controller/LoginControllerTest.java
@@ -1,0 +1,28 @@
+package hello;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.test.web.servlet.MockMvc;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import hello.ApplicationTest;
+
+@AutoConfigureMockMvc
+public class LoginControllerTest extends ApplicationTest
+{
+    private static final Logger LOG = LoggerFactory.getLogger(LoginControllerTest.class);
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    public void shouldReturnDefaultMessage() throws Exception {
+        this.mockMvc.perform(get("/login")).andDo(print()).andExpect(status().isOk());
+    }
+}
